+++ docs/source/changelog.rst
@@ -24,6 +24,8 @@ Core
 ++++
 
 -  Allow tuples as sharedict keys (:pr:`2763`)
+-  Calling compute within a dask.distributed task defaults to distributed
+   scheduler (:pr:`2762`)
 
 0.15.4 / 2017-10-06
 -------------------